fix(api): review fixes
