+++ packages/api/src/dataProviders/db/uniteLegalName/UniteLegalName.adapter.ts
@@ -23,4 +23,17 @@ export default class UniteLegalNameAdapter {
             _id: new ObjectId(entity.id),
         };
     }
+
+    static buildSearchKey(siren: Siren, name: string) {
+        const nameLc = name.toLowerCase();
+        let key = `${siren.value} - ${nameLc}`;
+        const removeAccents = str => str.normalize("NFD").replace(/[\u0300-\u036f]/g, ""); // Remove accent on name for futur search
+        const nameWithoutAccent = removeAccents(nameLc);
+
+        if (nameLc != nameWithoutAccent) {
+            key += ` - ${nameWithoutAccent}`;
+        }
+
+        return key;
+    }
 }
